Rename parameters module
Use a less generic name just in case IPA decides to add a parameters
plugin in the future.
